+++ src/device/mi.rs
@@ -1,9 +1,9 @@
 use crate::device;
 
-pub const MI_INIT_MODE_REG: u32 = 0;
-const MI_VERSION_REG: u32 = 1;
-pub const MI_INTR_REG: u32 = 2;
-pub const MI_INTR_MASK_REG: u32 = 3;
+pub const MI_INIT_MODE_REG: usize = 0;
+const MI_VERSION_REG: usize = 1;
+pub const MI_INTR_REG: usize = 2;
+pub const MI_INTR_MASK_REG: usize = 3;
 pub const MI_REGS_COUNT: usize = 4;
 
 /* read */
@@ -60,46 +60,45 @@ pub fn read_regs(
 
 pub fn write_regs(device: &mut device::Device, address: u64, value: u32, mask: u32) {
     let reg = (address & 0xFFFF) >> 2;
-    match reg as u32 {
+    match reg as usize {
         MI_INIT_MODE_REG => update_init_mode(device, value),
         MI_INTR_MASK_REG => update_intr_mask(device, value),
         _ => device::memory::masked_write_32(&mut device.mi.regs[reg as usize], value, mask),
     }
 
-    if device.mi.regs[MI_INTR_REG as usize] & device.mi.regs[MI_INTR_MASK_REG as usize] == 0 {
-        device.cpu.cop0.regs[device::cop0::COP0_CAUSE_REG as usize] &=
-            !device::cop0::COP0_CAUSE_IP2;
+    if device.mi.regs[MI_INTR_REG] & device.mi.regs[MI_INTR_MASK_REG] == 0 {
+        device.cpu.cop0.regs[device::cop0::COP0_CAUSE_REG] &= !device::cop0::COP0_CAUSE_IP2;
     }
     device::exceptions::check_pending_interrupts(device)
 }
 
 fn update_init_mode(device: &mut device::Device, w: u32) {
-    device.mi.regs[MI_INIT_MODE_REG as usize] &= !MI_INIT_LENGTH_MASK;
-    device.mi.regs[MI_INIT_MODE_REG as usize] |= w & MI_INIT_LENGTH_MASK;
+    device.mi.regs[MI_INIT_MODE_REG] &= !MI_INIT_LENGTH_MASK;
+    device.mi.regs[MI_INIT_MODE_REG] |= w & MI_INIT_LENGTH_MASK;
 
     if w & MI_CLR_INIT != 0 {
         for i in 0..(0x3F00000 >> 16) {
             device.memory.memory_map_write[i] = device::rdram::write_mem;
         }
-        device.mi.regs[MI_INIT_MODE_REG as usize] &= !MI_INIT_MODE
+        device.mi.regs[MI_INIT_MODE_REG] &= !MI_INIT_MODE
     }
     if w & MI_SET_INIT != 0 {
         for i in 0..(0x3F00000 >> 16) {
             device.memory.memory_map_write[i] = device::rdram::write_mem_repeat;
         }
-        device.mi.regs[MI_INIT_MODE_REG as usize] |= MI_INIT_MODE
+        device.mi.regs[MI_INIT_MODE_REG] |= MI_INIT_MODE
     }
     if w & MI_CLR_EBUS != 0 {
-        device.mi.regs[MI_INIT_MODE_REG as usize] &= !MI_EBUS_MODE
+        device.mi.regs[MI_INIT_MODE_REG] &= !MI_EBUS_MODE
     }
     if w & MI_SET_EBUS != 0 {
-        device.mi.regs[MI_INIT_MODE_REG as usize] |= MI_EBUS_MODE
+        device.mi.regs[MI_INIT_MODE_REG] |= MI_EBUS_MODE
     }
     if w & MI_CLR_RDRAM != 0 {
-        device.mi.regs[MI_INIT_MODE_REG as usize] &= !MI_RDRAM_MODE
+        device.mi.regs[MI_INIT_MODE_REG] &= !MI_RDRAM_MODE
     }
     if w & MI_SET_RDRAM != 0 {
-        device.mi.regs[MI_INIT_MODE_REG as usize] |= MI_RDRAM_MODE
+        device.mi.regs[MI_INIT_MODE_REG] |= MI_RDRAM_MODE
     }
 
     if w & MI_CLR_DP_INTR != 0 {
@@ -109,57 +108,56 @@ fn update_init_mode(device: &mut device::Device, w: u32) {
 
 fn update_intr_mask(device: &mut device::Device, w: u32) {
     if w & MI_CLR_SP != 0 {
-        device.mi.regs[MI_INTR_MASK_REG as usize] &= !MI_INTR_SP
+        device.mi.regs[MI_INTR_MASK_REG] &= !MI_INTR_SP
     }
     if w & MI_SET_SP != 0 {
-        device.mi.regs[MI_INTR_MASK_REG as usize] |= MI_INTR_SP
+        device.mi.regs[MI_INTR_MASK_REG] |= MI_INTR_SP
     }
     if w & MI_CLR_SI != 0 {
-        device.mi.regs[MI_INTR_MASK_REG as usize] &= !MI_INTR_SI
+        device.mi.regs[MI_INTR_MASK_REG] &= !MI_INTR_SI
     }
     if w & MI_SET_SI != 0 {
-        device.mi.regs[MI_INTR_MASK_REG as usize] |= MI_INTR_SI
+        device.mi.regs[MI_INTR_MASK_REG] |= MI_INTR_SI
     }
     if w & MI_CLR_AI != 0 {
-        device.mi.regs[MI_INTR_MASK_REG as usize] &= !MI_INTR_AI
+        device.mi.regs[MI_INTR_MASK_REG] &= !MI_INTR_AI
     }
     if w & MI_SET_AI != 0 {
-        device.mi.regs[MI_INTR_MASK_REG as usize] |= MI_INTR_AI
+        device.mi.regs[MI_INTR_MASK_REG] |= MI_INTR_AI
     }
     if w & MI_CLR_VI != 0 {
-        device.mi.regs[MI_INTR_MASK_REG as usize] &= !MI_INTR_VI
+        device.mi.regs[MI_INTR_MASK_REG] &= !MI_INTR_VI
     }
     if w & MI_SET_VI != 0 {
-        device.mi.regs[MI_INTR_MASK_REG as usize] |= MI_INTR_VI
+        device.mi.regs[MI_INTR_MASK_REG] |= MI_INTR_VI
     }
     if w & MI_CLR_PI != 0 {
-        device.mi.regs[MI_INTR_MASK_REG as usize] &= !MI_INTR_PI
+        device.mi.regs[MI_INTR_MASK_REG] &= !MI_INTR_PI
     }
     if w & MI_SET_PI != 0 {
-        device.mi.regs[MI_INTR_MASK_REG as usize] |= MI_INTR_PI
+        device.mi.regs[MI_INTR_MASK_REG] |= MI_INTR_PI
     }
     if w & MI_CLR_DP != 0 {
-        device.mi.regs[MI_INTR_MASK_REG as usize] &= !MI_INTR_DP
+        device.mi.regs[MI_INTR_MASK_REG] &= !MI_INTR_DP
     }
     if w & MI_SET_DP != 0 {
-        device.mi.regs[MI_INTR_MASK_REG as usize] |= MI_INTR_DP
+        device.mi.regs[MI_INTR_MASK_REG] |= MI_INTR_DP
     }
 }
 
 pub fn clear_rcp_interrupt(device: &mut device::Device, interrupt: u32) {
-    device.mi.regs[MI_INTR_REG as usize] &= !interrupt;
+    device.mi.regs[MI_INTR_REG] &= !interrupt;
 
-    if device.mi.regs[MI_INTR_REG as usize] & device.mi.regs[MI_INTR_MASK_REG as usize] == 0 {
-        device.cpu.cop0.regs[device::cop0::COP0_CAUSE_REG as usize] &=
-            !device::cop0::COP0_CAUSE_IP2;
+    if device.mi.regs[MI_INTR_REG] & device.mi.regs[MI_INTR_MASK_REG] == 0 {
+        device.cpu.cop0.regs[device::cop0::COP0_CAUSE_REG] &= !device::cop0::COP0_CAUSE_IP2;
     }
 }
 
 pub fn set_rcp_interrupt(device: &mut device::Device, interrupt: u32) {
-    device.mi.regs[MI_INTR_REG as usize] |= interrupt;
+    device.mi.regs[MI_INTR_REG] |= interrupt;
     device::exceptions::check_pending_interrupts(device)
 }
 
 pub fn init(device: &mut device::Device) {
-    device.mi.regs[MI_VERSION_REG as usize] = 0x02020102
+    device.mi.regs[MI_VERSION_REG] = 0x02020102
 }
