+++ src/device/pif.rs
@@ -68,7 +68,7 @@ pub fn write_mem(device: &mut device::Device, address: u64, value: u32, mask: u3
 
     device.si.dma_dir = device::si::DmaDir::Write;
     device::events::create_event(device, device::events::EVENT_TYPE_SI, 3200); //based on https://github.com/rasky/n64-systembench
-    device.si.regs[device::si::SI_STATUS_REG as usize] |=
+    device.si.regs[device::si::SI_STATUS_REG] |=
         device::si::SI_STATUS_DMA_BUSY | device::si::SI_STATUS_IO_BUSY
 }
